Move around version tests
